requirements.txt
pocketflow==0.0.2
openai==1.82.0
anthropic>=0.5.0
python-dotenv==1.0.1
requests>=2.31.0
pydantic>=2.0.0
pyyaml==6.0.1
rich==13.7.0 
firecrawl-py==2.7.1
google-genai==1.17.0
sqlalchemy>=2.0.41
psycopg2-binary>=2.9.10
pinecone>=7.0.2
alembic>=1.16.1
matplotlib==3.10.3
seaborn==0.13.2
numpy==2.2.6
supabase==2.15.2
